+++ healthchecks/src/main/java/uk/gov/justice/services/healthcheck/healthchecks/artemis/DestinationNotFoundException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.healthcheck.healthchecks.artemis;
+
+public class DestinationNotFoundException extends Exception {
+
+    public DestinationNotFoundException(String message, Exception cause) {
+        super(message, cause);
+    }
+}